+++ core/src/server/delivery/request/BaseRequest.cpp
@@ -11,13 +11,72 @@
 
 #include "server/delivery/request/BaseRequest.h"
 #include "utils/CommonUtil.h"
+#include "utils/Exception.h"
 #include "utils/Log.h"
+
+#include <map>
 
 namespace milvus {
 namespace server {
 
-BaseRequest::BaseRequest(const std::shared_ptr<Context>& context, const std::string& request_group, bool async)
-    : context_(context), request_group_(request_group), async_(async), done_(false) {
+static const char* DQL_REQUEST_GROUP = "dql";
+static const char* DDL_DML_REQUEST_GROUP = "ddl_dml";
+static const char* INFO_REQUEST_GROUP = "info";
+
+namespace {
+std::string
+RequestGroup(BaseRequest::RequestType type) {
+    static std::map<BaseRequest::RequestType, std::string> s_map_type_group = {
+        // general operations
+        {BaseRequest::kCmd, INFO_REQUEST_GROUP},
+
+        // data operations
+        {BaseRequest::kInsert, DDL_DML_REQUEST_GROUP},
+        {BaseRequest::kCompact, DDL_DML_REQUEST_GROUP},
+        {BaseRequest::kFlush, DDL_DML_REQUEST_GROUP},
+        {BaseRequest::kDeleteByID, DDL_DML_REQUEST_GROUP},
+        {BaseRequest::kGetVectorByID, INFO_REQUEST_GROUP},
+        {BaseRequest::kGetVectorIDs, INFO_REQUEST_GROUP},
+
+        // table operations
+        {BaseRequest::kShowTables, INFO_REQUEST_GROUP},
+        {BaseRequest::kCreateTable, DDL_DML_REQUEST_GROUP},
+        {BaseRequest::kHasTable, INFO_REQUEST_GROUP},
+        {BaseRequest::kDescribeTable, INFO_REQUEST_GROUP},
+        {BaseRequest::kCountTable, INFO_REQUEST_GROUP},
+        {BaseRequest::kShowTableInfo, INFO_REQUEST_GROUP},
+        {BaseRequest::kDropTable, DDL_DML_REQUEST_GROUP},
+        {BaseRequest::kPreloadTable, DQL_REQUEST_GROUP},
+
+        // partition operations
+        {BaseRequest::kCreatePartition, DDL_DML_REQUEST_GROUP},
+        {BaseRequest::kShowPartitions, INFO_REQUEST_GROUP},
+        {BaseRequest::kDropPartition, DDL_DML_REQUEST_GROUP},
+
+        // index operations
+        {BaseRequest::kCreateIndex, DDL_DML_REQUEST_GROUP},
+        {BaseRequest::kDescribeIndex, INFO_REQUEST_GROUP},
+        {BaseRequest::kDropIndex, DDL_DML_REQUEST_GROUP},
+
+        // search operations
+        {BaseRequest::kSearchByID, DQL_REQUEST_GROUP},
+        {BaseRequest::kSearch, DQL_REQUEST_GROUP},
+        {BaseRequest::kSearchCombine, DQL_REQUEST_GROUP},
+    };
+
+    auto iter = s_map_type_group.find(type);
+    if (iter == s_map_type_group.end()) {
+        SERVER_LOG_ERROR << "Unsupported request type: " << type;
+        throw Exception(SERVER_NOT_IMPLEMENT, "request group undefined");
+    }
+    return iter->second;
+}
+}  // namespace
+
+BaseRequest::BaseRequest(const std::shared_ptr<milvus::server::Context>& context, BaseRequest::RequestType type,
+                         bool async)
+    : context_(context), type_(type), async_(async), done_(false) {
+    request_group_ = milvus::server::RequestGroup(type);
 }
 
 BaseRequest::~BaseRequest() {
@@ -37,11 +96,12 @@ BaseRequest::Done() {
     finish_cond_.notify_all();
 }
 
-Status
-BaseRequest::SetStatus(ErrorCode error_code, const std::string& error_msg) {
-    status_ = Status(error_code, error_msg);
-    SERVER_LOG_ERROR << error_msg;
-    return status_;
+void
+BaseRequest::set_status(const Status& status) {
+    status_ = status;
+    if (!status_.ok()) {
+        SERVER_LOG_ERROR << status_.message();
+    }
 }
 
 std::string
@@ -55,7 +115,6 @@ Status
 BaseRequest::WaitToFinish() {
     std::unique_lock<std::mutex> lock(finish_mtx_);
     finish_cond_.wait(lock, [this] { return done_; });
-
     return status_;
 }