+++ core/src/server/delivery/request/HasTableRequest.h
@@ -22,10 +22,11 @@ namespace server {
 class HasTableRequest : public BaseRequest {
  public:
     static BaseRequestPtr
-    Create(const std::shared_ptr<Context>& context, const std::string& table_name, bool& has_table);
+    Create(const std::shared_ptr<milvus::server::Context>& context, const std::string& table_name, bool& has_table);
 
  protected:
-    HasTableRequest(const std::shared_ptr<Context>& context, const std::string& table_name, bool& has_table);
+    HasTableRequest(const std::shared_ptr<milvus::server::Context>& context, const std::string& table_name,
+                    bool& has_table);
 
     Status
     OnExecute() override;